Split client_test.py into package
